Create Suspicious MSHTA Usage.kql
I have found in the various environments I have been in that mshta in the modern world is far and few between. More often than not, it is used for malicious activity over odd edge cases. Edge cases can be tuned out and this could become an effective detection for most environments.